Widgets
Added a widgets table similar in format to tags. It defines the
widget layouts for both laptops and desktops and assigns
mywidgets.widgets to the correct one based on the number of screens
present. The for_each_screen function then adds them to the wibars
based on which screen is being initialized.