Buttons Error
When trying to use the docker widget, it would error out when trying
to add mouse buttons to the widget. This looks to be some sort of
scoping issue, likely due to the version of lua currently being
used (5.4.6). Regardless of the issue, moving the widget definition
from the global level to within the worker function fixes the
issue. This seems to be a common design pattern in other widgets.